conversion functions return data in "natural" shape when possible
For source data such as unboxed vector, where we have size information alread, we should use this information rather than requiring the user to supply a target shape and checking that these two match. This means that we always return as a 1D array, but the user can later reshape this for free so not so problematic (in 1.2 there will be a haskell-level reshape operation as well).